+++ tests/Functional/Elastica/ElasticaRulerTest.php
@@ -14,6 +14,7 @@ declare(strict_types = 1);
 namespace FiveLab\Component\Ruler\Tests\Functional\Elastica;
 
 use Elastica\Query;
+use FiveLab\Component\Ruler\Query\RawSearchQuery;
 use FiveLab\Component\Ruler\Ruler;
 use FiveLab\Component\Ruler\Target\ElasticaTarget;
 use PHPUnit\Framework\TestCase;
@@ -37,18 +38,17 @@ class ElasticaRulerTest extends TestCase
      * @test
      *
      * @param string $file
+     * @param mixed  $query
      *
      * @dataProvider provideDataForApply
      */
-    public function shouldSuccessApply(string $file): void
+    public function shouldSuccessApply(string $file, mixed $query): void
     {
         $data = \json_decode(\file_get_contents($file), true, 512, JSON_THROW_ON_ERROR);
 
         $rule = $data['rule'];
         $params = $data['params'];
         $expectedQuery = $data['query'];
-
-        $query = new Query();
 
         $this->ruler->apply($query, $rule, $params);
 
@@ -57,26 +57,34 @@ class ElasticaRulerTest extends TestCase
 
     /**
      * @test
+     *
+     * @param mixed $query
+     *
+     * @dataProvider provideQuery
      */
-    public function shouldFailForMoreNested(): void
+    public function shouldFailForMoreNested(mixed $query): void
     {
         $this->expectException(\RuntimeException::class);
         $this->expectExceptionMessage('Only one nested level supported.');
 
-        $this->ruler->apply(new Query(), 'products.variants.categories.key = :key', [
+        $this->ruler->apply($query, 'products.variants.categories.key = :key', [
             'key' => 'foo',
         ]);
     }
 
     /**
      * @test
+     *
+     * @param mixed $query
+     *
+     * @dataProvider provideQuery
      */
-    public function shouldFailIfParameterMissed(): void
+    public function shouldFailIfParameterMissed(mixed $query): void
     {
         $this->expectException(\LogicException::class);
         $this->expectExceptionMessage('The parameter "id" is missed. Possible parameters are "foo", "some".');
 
-        $this->ruler->apply(new Query(), 'id = :id', [
+        $this->ruler->apply($query, 'id = :id', [
             'foo'  => 'bar',
             'some' => 1,
         ]);
@@ -89,20 +97,42 @@ class ElasticaRulerTest extends TestCase
      */
     public function provideDataForApply(): array
     {
+        $files = [
+            __DIR__.'/Resources/eq.json',
+            __DIR__.'/Resources/not-eq.json',
+            __DIR__.'/Resources/in.json',
+            __DIR__.'/Resources/not-in.json',
+            __DIR__.'/Resources/gt.json',
+            __DIR__.'/Resources/gte.json',
+            __DIR__.'/Resources/lt.json',
+            __DIR__.'/Resources/lte.json',
+            __DIR__.'/Resources/like.json',
+            __DIR__.'/Resources/and.json',
+            __DIR__.'/Resources/or.json',
+            __DIR__.'/Resources/nested.json',
+            __DIR__.'/Resources/combined-logical.json',
+        ];
+
+        return \array_merge(
+            \array_map(static function (string $file): array {
+                return [$file, new Query()];
+            }, $files),
+            \array_map(static function (string $file): array {
+                return [$file, new RawSearchQuery()];
+            }, $files)
+        );
+    }
+
+    /**
+     * Provide query
+     *
+     * @return array
+     */
+    public function provideQuery(): array
+    {
         return [
-            [__DIR__.'/Resources/eq.json'],
-            [__DIR__.'/Resources/not-eq.json'],
-            [__DIR__.'/Resources/in.json'],
-            [__DIR__.'/Resources/not-in.json'],
-            [__DIR__.'/Resources/gt.json'],
-            [__DIR__.'/Resources/gte.json'],
-            [__DIR__.'/Resources/lt.json'],
-            [__DIR__.'/Resources/lte.json'],
-            [__DIR__.'/Resources/like.json'],
-            [__DIR__.'/Resources/and.json'],
-            [__DIR__.'/Resources/or.json'],
-            [__DIR__.'/Resources/nested.json'],
-            [__DIR__.'/Resources/combined-logical.json'],
+            [new Query()],
+            [new RawSearchQuery()],
         ];
     }
 }