+++ tests/Unit/Target/ElasticaTargetTest.php
@@ -14,6 +14,7 @@ declare(strict_types = 1);
 namespace FiveLab\Component\Ruler\Unit\Elastica;
 
 use Elastica\Query;
+use FiveLab\Component\Ruler\Query\RawSearchQuery;
 use FiveLab\Component\Ruler\Target\ElasticaTarget;
 use PHPUnit\Framework\TestCase;
 
@@ -34,10 +35,14 @@ class ElasticaTargetTest extends TestCase
 
     /**
      * @test
+     *
+     * @param mixed $query
+     *
+     * @dataProvider provideQuery
      */
-    public function shouldSuccessSupports(): void
+    public function shouldSuccessSupports(mixed $query): void
     {
-        $supports = $this->target->supports(new Query());
+        $supports = $this->target->supports($query);
 
         self::assertTrue($supports);
     }
@@ -50,5 +55,18 @@ class ElasticaTargetTest extends TestCase
         $supports = $this->target->supports(new \stdClass());
 
         self::assertFalse($supports);
+    }
+
+    /**
+     * Provide query
+     *
+     * @return array
+     */
+    public function provideQuery(): array
+    {
+        return [
+            [new Query()],
+            [new RawSearchQuery()],
+        ];
     }
 }